RD-4279: Renamed TerraformModal directory
