+++ widgets/common/src/terraformModal/images/terraform-icon.png
